refactor(cheat-detection)!: rename the detection package to cheat_detection
"detection" says nothing about what is being detected, and it was the one
surface still saying it: the record field is already cheating_report, the
toggles are BENCH_CHEAT_DETECT / BENCH_CHEAT_INVENTORY, and the docs page
is titled "Cheating detection". The directory now matches.

Pure rename -- devops_bench/detection -> devops_bench/cheat_detection,
tests/unit/detection -> tests/unit/cheat_detection, and
docs/components/detection.md -> cheat-detection.md -- with references
rewritten. No behaviour change.

Naming note for reviewers: "contamination" is the term of art in the ML
benchmark literature but means training-set leakage, not an agent reading
answers at runtime, so it would mislead rather than clarify.
