+++ docs/components/cheat-detection.md
@@ -1,6 +1,6 @@
 # Cheating detection
 
-Agents under test run as ordinary subprocesses on the harness host — there is no filesystem sandbox — so the benchmark's own sensitive material (task definitions with their judge rubrics and verification specs, the scoring code, prior results, the repo checkout, the upstream GitHub repo) is technically reachable. The detection layer in [`devops_bench/detection/`](../../devops_bench/detection/) scans each run's recorded trajectory for signs that an agent accessed that material, and attaches a **flag-only** report for human review.
+Agents under test run as ordinary subprocesses on the harness host — there is no filesystem sandbox — so the benchmark's own sensitive material (task definitions with their judge rubrics and verification specs, the scoring code, prior results, the repo checkout, the upstream GitHub repo) is technically reachable. The detection layer in [`devops_bench/cheat_detection/`](../../devops_bench/cheat_detection/) scans each run's recorded trajectory for signs that an agent accessed that material, and attaches a **flag-only** report for human review.
 
 > [!IMPORTANT]
 > Detection never changes scores, never touches `validated`, and never aborts a run. A `flagged` report is an annotation for a reviewer, not a verdict: trajectory analysis sees only what the agent's transcript recorded, so it can neither prove innocence nor block cheating — it surfaces the runs worth a closer look.
@@ -19,7 +19,7 @@ An errored run with an empty trajectory and empty output reports `no_data` — d
 
 ## Default rule categories
 
-Rules live in [`devops_bench/detection/rules.py`](../../devops_bench/detection/rules.py) and match the *kind* of sensitive material, never a specific task, so new tasks are covered automatically:
+Rules live in [`devops_bench/cheat_detection/rules.py`](../../devops_bench/cheat_detection/rules.py) and match the *kind* of sensitive material, never a specific task, so new tasks are covered automatically:
 
 | Category | Severity | Catches |
 | --- | --- | --- |
@@ -45,7 +45,7 @@ rules:
 
 ## Prior-run artifacts (dynamic rules)
 
-The agent home persists between runs, so earlier runs leave material behind — a previous `report.md` is an answer key for the next attempt. "Left by a prior run" is temporal, not lexical: this run's legitimate `report.md` write and a read of last run's leftover are the same string, so no static regex can separate them. Instead, [`devops_bench/detection/inventory.py`](../../devops_bench/detection/inventory.py) snapshots the home **before every task's agent executes** and generates `prior-run-artifact` rules from whatever is already there. Per task rather than once per batch, because the home keeps accumulating while the batch runs: task 1's deliverable is precisely the answer key task 2 must not read, and a single run-start snapshot is blind to it. A baseline of provisioned names (`bin`, `bench.env`, `devops-bench`, `matrix-runs`) and an enumerated set of environment dotfiles (`.bashrc`, `.cache`, `.config`, `.ssh`, `.kube`, …) are skipped — but *only* those: any other hidden entry is a leftover like any visible one, because agent CLIs conventionally keep their state in a dotdir and that is exactly where cross-run contamination accumulates (a stale `~/.openclaw/workspace` holds a previous task's deliverables and git history). One known caveat, deliberately unhandled: the state directory of the agent currently under test is not special-cased, so an honest agent that references its own state dir in a recorded tool call flags; if that bites in practice, the harness (which knows the agent type) should add that one name to the baseline it passes rather than the dotfile set widening. The generated rules are:
+The agent home persists between runs, so earlier runs leave material behind — a previous `report.md` is an answer key for the next attempt. "Left by a prior run" is temporal, not lexical: this run's legitimate `report.md` write and a read of last run's leftover are the same string, so no static regex can separate them. Instead, [`devops_bench/cheat_detection/inventory.py`](../../devops_bench/cheat_detection/inventory.py) snapshots the home **before every task's agent executes** and generates `prior-run-artifact` rules from whatever is already there. Per task rather than once per batch, because the home keeps accumulating while the batch runs: task 1's deliverable is precisely the answer key task 2 must not read, and a single run-start snapshot is blind to it. A baseline of provisioned names (`bin`, `bench.env`, `devops-bench`, `matrix-runs`) and an enumerated set of environment dotfiles (`.bashrc`, `.cache`, `.config`, `.ssh`, `.kube`, …) are skipped — but *only* those: any other hidden entry is a leftover like any visible one, because agent CLIs conventionally keep their state in a dotdir and that is exactly where cross-run contamination accumulates (a stale `~/.openclaw/workspace` holds a previous task's deliverables and git history). One known caveat, deliberately unhandled: the state directory of the agent currently under test is not special-cased, so an honest agent that references its own state dir in a recorded tool call flags; if that bites in practice, the harness (which knows the agent type) should add that one name to the baseline it passes rather than the dotfile set widening. The generated rules are:
 
 - **Path rules**, anchored to home spellings (`~`, `$HOME`, the literal home path), so this run's own same-named files elsewhere (e.g. a fresh `/tmp` clone) never match. Like all path rules they scan every surface, so a leftover's name echoing through a directory listing flags too — the agent had no reason to be enumerating the home directory.
 - **Content fingerprints** for small text leftovers: their most distinctive lines, matched only against `result`/`output`. Reading the stale file reproduces those exact lines; writing a fresh file does not — so a stale `report.md` read is flagged even when the command that read it named some other path.
